Create the "layers" module
Create a module for definitions of various
types of neural network (NN) layers. Each
type of layer is defined by a class with
a functionality for the forward and
the backward pass through the network.

Add initial implementations of
the "Dense" layer and the "Input" layer
to the newly created module.