Reimplement the "activations" module
Implement the new format of the activations functions.
Every activation function can now be created by creating
the object of the ActivationFunction class. The constructor
accepts a string name of the function and then dynamically
selects a callable function from the already defined functions
in the module. The object itself is a callable object and can
be directly used to compute the output of the function - eg.:

f = ActivationFunction("relu")
x = np.random.randn(4, 1)
result = f(x)