Reimplement the "losses" module
Implement the new format of the loss functions.
Every loss function can now be created by creating
the object of the LossFunction class. The constructor
accepts a string name of the function and then dynamically
selects a callable function from the already defined functions
in the module. The object itself is a callable object and can
be directly used to compute the output of the function - eg.:

f = LossFunction("negative_log_loss")
y_hat = [0, 0.7]
y = [0, 1]
result = f(y_hat, y)